refactor: split app route components
