+++ src/components/editor.tsx
@@ -1,0 +1,122 @@
+import {
+  CircleHelpIcon,
+  SettingsIcon,
+  SlidersHorizontalIcon,
+} from "lucide-react";
+import { useEffect, useState } from "react";
+import { useWindowEvent } from "../hooks/use-window-event";
+import { isShortcutTextInputTarget, matchKeyboardEvent } from "../lib/keyboard";
+import { projectStorage } from "../lib/project-storage";
+import { HelpOverlay } from "./help-overlay";
+import { Mixer } from "./mixer";
+import { PianoRoll } from "./piano-roll";
+import { Settings } from "./settings";
+import { Transport } from "./transport";
+import { Button } from "./ui/button";
+import { Dialog } from "./ui/dialog";
+
+type EditorProps = {
+  projectId: string;
+  initialProjectName: string;
+};
+
+export function Editor({ projectId, initialProjectName }: EditorProps) {
+  const [isHelpOpen, setIsHelpOpen] = useState(false);
+  const [isSettingsOpen, setIsSettingsOpen] = useState(false);
+  const [isMixerOpen, setIsMixerOpen] = useState(false);
+  const [projectName, setProjectName] = useState(initialProjectName);
+
+  // Update document title when project name changes
+  useEffect(() => {
+    document.title = `${projectName} - Toy MIDI`;
+  }, [projectName]);
+
+  // Keyboard shortcuts for overlays
+  useWindowEvent("keydown", (e) => {
+    if (isShortcutTextInputTarget(e.target)) {
+      return;
+    }
+
+    if (matchKeyboardEvent(e, "Escape")) {
+      if (isSettingsOpen) {
+        e.preventDefault();
+        e.stopPropagation();
+        setIsSettingsOpen(false);
+      } else if (isMixerOpen) {
+        e.preventDefault();
+        e.stopPropagation();
+        setIsMixerOpen(false);
+      } else if (isHelpOpen) {
+        e.preventDefault();
+        e.stopPropagation();
+        setIsHelpOpen(false);
+      }
+    }
+    if (e.key === "?" && !e.repeat) {
+      e.preventDefault();
+      setIsHelpOpen((prev) => !prev);
+    }
+  });
+
+  return (
+    <div className="h-screen flex flex-col bg-neutral-900">
+      <Transport
+        projectName={projectName}
+        controls={
+          <>
+            <Button
+              data-testid="settings-button"
+              onClick={() => setIsSettingsOpen(true)}
+              title="Settings"
+              className="size-9 hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50"
+            >
+              <SettingsIcon className="size-5" />
+            </Button>
+            <Button
+              data-testid="mixer-button"
+              onClick={() => setIsMixerOpen(true)}
+              title="Mixer"
+              className="size-9 hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50"
+            >
+              <SlidersHorizontalIcon className="size-5" />
+            </Button>
+            <Button
+              data-testid="help-button"
+              onClick={() => setIsHelpOpen(true)}
+              title="Show keyboard shortcuts (?)"
+              className="size-9 hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50"
+            >
+              <CircleHelpIcon className="size-5" />
+            </Button>
+          </>
+        }
+      />
+      <PianoRoll />
+      <HelpOverlay isOpen={isHelpOpen} onClose={() => setIsHelpOpen(false)} />
+      <Dialog
+        isOpen={isMixerOpen}
+        onClose={() => setIsMixerOpen(false)}
+        title="Mixer"
+        testId="mixer-dialog"
+      >
+        <Mixer />
+      </Dialog>
+      <Dialog
+        isOpen={isSettingsOpen}
+        onClose={() => setIsSettingsOpen(false)}
+        title="Settings"
+        testId="settings-dialog"
+      >
+        <Settings
+          projectName={projectName}
+          onProjectNameChange={(name) => {
+            if (name && name !== projectName) {
+              projectStorage.updateMetadata(projectId, { name });
+              setProjectName(name);
+            }
+          }}
+        />
+      </Dialog>
+    </div>
+  );
+}
